mostly generating

+++ vitact/vitact/tardataset.py
@@ -4,12 +4,18 @@ import os
 import time
 import glob
 import webdataset as wds
+from PIL import Image
+import io
+
+from torchvision.io import decode_jpeg, ImageReadMode
+import torchvision.transforms as transforms
 
 class StreamingDataset(IterableDataset):
-    def __init__(self, data_dir):
+    def __init__(self, data_dir, destructive=False):
         self.data_dir = data_dir
         self.stop = False
         self.exclude = set()
+        self.destructive = destructive
 
     def _get_tar_files(self):
         """Retrieve the list of .ready.tar files in the data directory."""
@@ -35,5 +41,29 @@ class StreamingDataset(IterableDataset):
                 for sample in dataset:
                     yield sample
                 
-                os.remove(tar_file)
+                if self.destructive:
+                    os.remove(tar_file)
+
+class StreamingTensorDataset(StreamingDataset):
+    def __init__(self, data_dir, destructive=False, device='cuda'):
+        super().__init__(data_dir, destructive)
+
+        self.transform = transforms.ToTensor()
+        self.device = device
+
+    def __iter__(self):
+        for sample in super().__iter__():
+            if self.stop:
+                break
+
+            if 'jpg' in sample:
+                try:
+                    image = Image.open(io.BytesIO(sample['jpg'])).convert('RGB')
+                    image = image.resize((224, 224))
+
+                    image_tensor = self.transform(image).to(self.device)
+                    yield image_tensor
+                except Exception as e:
+                    print(type(e), e)
+
